PythonBridge/requirements_minimal.txt
# Minimal requirements for MFLUX Mac App
# Core dependencies only - for production use

# Core MLX and MFLUX
mlx>=0.2.0
mflux-ai>=0.1.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Basic utilities
requests>=2.31.0 